+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 ## Version 0.6.4-dev
 
 ### Bug Fixes
+
 * Fixed a bug in the documentation in which the `source` button would link to decorator code, instead of the relevant function ([#2184](https://github.com/scikit-bio/scikit-bio/pull/2184)).
+
+### Miscellaneous
+
+* Updated documentation to include description of how to stream data through stdin with scikit-bio's `read` function ([2185](https://github.com/scikit-bio/scikit-bio/pull/2185))
 
 
 ## Version 0.6.3